+++ plugins/nf-quilt/src/main/nextflow/quilt/QuiltObserver.groovy
@@ -25,7 +25,6 @@ import java.nio.file.Path
 import java.nio.file.Paths
 import java.time.LocalDateTime
 
-import groovy.json.JsonOutput
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import nextflow.Session
@@ -83,23 +82,6 @@ class QuiltObserver implements TraceObserver {
         }
         return null
     }
-
-    static String toJson(Map dict) {
-        List<String> entries = dict.collect { key, value ->
-            String prefix = JsonOutput.toJson(key)
-            String suffix = "Cannot generate JSON for: ${value}"
-            log.info("QuiltObserver.toJson: ${prefix} [${suffix.length()}]")
-            try {
-                suffix = JsonOutput.toJson(value)
-            }
-            catch (Exception e) {
-                log.error(suffix, e)
-            }
-            return "${prefix}:${suffix}".toString()
-        }
-        return "{${entries.join(',')}}".toString()
-    }
-
     @Override
     void onFlowCreate(Session session) {
         log.debug("`onFlowCreate` $this")
@@ -151,20 +133,18 @@ class QuiltObserver implements TraceObserver {
         String msg = pkg
         Map meta = [pkg: msg]
         String text = 'Stub README'
-        String jsonMeta = JsonOutput.toJson(meta)
         try {
             meta = getMetadata()
             msg = "${meta['config']['runName']}: ${meta['cmd']}"
             text = readme(meta, msg)
             meta.remove('config')
-            jsonMeta = QuiltObserver.toJson(meta)
         }
         catch (Exception e) {
             log.error('publish: cannot generate metadata (QuiltObserver uninitialized?)', e)
         }
         writeString(text, pkg, 'README.md')
         writeString("$meta", pkg, 'quilt_metadata.txt')
-        def rc = pkg.push(msg, jsonMeta)
+        def rc = pkg.push(msg, meta)
         log.info("$rc: pushed package[$pkg] $msg")
     }
